NetBSD compile fixes

+++ src/common/stringbuffer.h
@@ -37,7 +37,7 @@ public:
     template<class T>
     StringBuffer& operator<<(const T& v)
     {
-        if(buf_.tellp() != 0)
+        if(ssize_t(buf_.tellp()) != 0)
             buf_ << ' ';
 
         buf_ << v;
